Fix a lot of compiler warnings.
git-svn-id: svn://svn.code.sf.net/p/boomerang/code/trunk@2464 dc5e1367-85d4-4bd1-96b4-a20bfbfd92d5

+++ boomerang/type/type.cpp
@@ -89,7 +89,7 @@ void PointerType::setPointsTo(Type* p)
       // Note: comparing pointers
       points_to = new VoidType();		// Can't point to self; impossible to compare, print, etc
       if (VERBOSE)
-        LOG << "Warning: attempted to create pointer to self: " << this << "\n";
+        LOG << "Warning: attempted to create pointer to self: " << (unsigned int)this << "\n";
     }
   else
     points_to = p;
